+++ lbrynet/__init__.py
@@ -1,6 +1,6 @@
 import logging
 
-__version__ = "0.9.2rc24"
+__version__ = "0.10.0rc1"
 version = tuple(__version__.split('.'))
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())